requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and ORM
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Redis for caching
redis==5.0.1
hiredis==2.2.3

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Task processing
celery==5.3.4
flower==2.0.1

# HTTP client
httpx==0.25.2
aiofiles==23.2.1

# Machine Learning and AI
torch==2.1.1
torchvision==0.16.1
torchaudio==0.16.1
transformers==4.35.2
numpy==1.24.4
scikit-learn==1.3.2

# Computer Vision
opencv-python==4.8.1.78
pillow==10.1.0

# Audio processing
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1

# Video processing
ffmpeg-python==0.2.0

# Data processing
pandas==2.1.3
matplotlib==3.8.2
seaborn==0.13.0

# Configuration and environment
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8

# Additional utilities
python-json-logger==2.0.7
email-validator==2.1.0
phonenumbers==8.13.25
